Move task helper logic to redux actions
