+++ packages/colony-starter-react/src/client/actions/taskActions.jsx
@@ -1,20 +1,44 @@
+import BN from 'bn.js'
+import { sha3 } from 'web3-utils';
 import { store } from '../index'
 import * as actions from '../constants/actions'
-import * as taskActions from '../../helpers/actions/taskActions'
 import { setStatePots } from './fundingActions'
+import ecp from '../../helpers/ecp'
+import ms from '../../helpers/ms'
 
 // cancelTask
 
 export const cancelTask = (colonyClient, taskId) => ({
   type: actions.CANCEL_TASK,
-  payload: taskActions.cancelTask(colonyClient, taskId)
-    .then(task => {
-      store.dispatch(setStateTask(task))
-      store.dispatch(cancelTaskSuccess(true))
+  payload: (async () => {
+
+    // Start multisignature operation
+    const operation = await colonyClient.cancelTask.startOperation({
+      taskId,
     })
-    .catch(error => {
-      store.dispatch(cancelTaskError(error.message))
-    }),
+
+    // Handle multisignature operation
+    const tx = await ms.signOperation(operation);
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(cancelTaskSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(cancelTaskError(error.message))
+  }),
 })
 
 export const cancelTaskError = (error) => ({
@@ -31,13 +55,34 @@ export const cancelTaskSuccess = (success) => ({
 
 export const claimPayout = (colonyClient, taskId, role) => ({
   type: actions.CLAIM_PAYOUT,
-  payload: taskActions.claimPayout(colonyClient, taskId, role)
-    .then(success => {
-      store.dispatch(claimPayoutSuccess(true))
+  payload: (async () => {
+
+    // Claim payout
+    const tx = await colonyClient.claimPayout.send({
+      taskId,
+      role,
+      token: colonyClient.tokenClient.contract.address,
     })
-    .catch(error => {
-      store.dispatch(claimPayoutError(error.message))
-    }),
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(claimPayoutSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(claimPayoutError(error.message))
+  }),
 })
 
 export const claimPayoutError = (error) => ({
@@ -54,15 +99,52 @@ export const claimPayoutSuccess = (success) => ({
 
 export const createTask = (colonyClient, task) => ({
   type: actions.CREATE_TASK,
-  payload: taskActions.createTask(colonyClient, task)
-    .then(taskId => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(setStateTaskCount(taskId))
-      store.dispatch(createTaskSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(createTaskError(error.message))
-    }),
+  payload: (async () => {
+
+    // Declare new task
+    const newTask = {}
+
+    // Initialize extended protocol
+    await ecp.init()
+
+    // Create specification hash
+    newTask.specificationHash = await ecp.saveTaskSpecification(task.specification)
+
+    // Stop extended protocol
+    await ecp.stop()
+
+    // Set domain id if provided
+    if (task.domainId) newTask.domainId = task.domainId
+
+    // Set skill id if provided
+    if (task.skillId) newTask.skillId = task.skillId
+
+    // Set due date if provided
+    if (task.dueDate) newTask.dueDate = new Date(task.dueDate)
+
+    // Create task
+    const tx = await colonyClient.createTask.send(newTask)
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return task id
+    return tx.eventData.taskId
+
+  })()
+  .then(taskId => {
+
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(createTaskSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(createTaskError(error.message))
+  }),
 })
 
 export const createTaskError = (error) => ({
@@ -79,14 +161,30 @@ export const createTaskSuccess = (success) => ({
 
 export const finalizeTask = (colonyClient, taskId) => ({
   type: actions.FINALIZE_TASK,
-  payload: taskActions.finalizeTask(colonyClient, taskId)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(finalizeTaskSuccess(success))
-    })
-    .catch(error => {
-      store.dispatch(finalizeTaskError(error.message))
-    }),
+  payload: (async () => {
+
+    // Finalize task
+    const tx = await colonyClient.finalizeTask.send({ taskId })
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(finalizeTaskSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(finalizeTaskError(error.message))
+  }),
 })
 
 export const finalizeTaskError = (error) => ({
@@ -103,15 +201,44 @@ export const finalizeTaskSuccess = (success) => ({
 
 export const fundTask = (colonyClient, taskId, amount) => ({
   type: actions.FUND_TASK,
-  payload: taskActions.fundTask(colonyClient, taskId, amount)
-    .then(success => {
-      store.dispatch(setStatePots(null))
-      store.dispatch(setStateTasks(null))
-      store.dispatch(fundTaskSuccess(true))
+  payload: (async () => {
+
+    // Get task
+    const task = await colonyClient.getTask.call({ taskId })
+
+    // Get task domain
+    const domain = await colonyClient.getDomain.call({
+      domainId: task.domainId,
     })
-    .catch(error => {
-      store.dispatch(fundTaskError(error.message))
-    }),
+
+    // Move funds between pots
+    const tx = await colonyClient.moveFundsBetweenPots.send({
+      fromPot: domain.potId,
+      toPot: task.potId,
+      amount: new BN(amount),
+      token: colonyClient.tokenClient.contract.address,
+    })
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(setStatePots(null))
+    store.dispatch(setStateTasks(null))
+    store.dispatch(fundTaskSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(fundTaskError(error.message))
+  }),
 })
 
 export const fundTaskError = (error) => ({
@@ -128,14 +255,22 @@ export const fundTaskSuccess = (success) => ({
 
 export const getMultisigOperations = (colonyClient, taskId) => ({
   type: actions.GET_MULTISIG_OPERATIONS,
-  payload: taskActions.getMultisigOperations(colonyClient, taskId)
-    .then(multisigOperations => {
-      store.dispatch(setStateMultisigOperations(multisigOperations))
-      store.dispatch(getMultisigOperationsSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(getMultisigOperationsError(error.message))
-    }),
+  payload: (async () => {
+
+    // Get operations
+    const multisigOperations = ms.getOperations(colonyClient, taskId)
+
+    // Return operations
+    return multisigOperations
+
+  })()
+  .then(multisigOperations => {
+    store.dispatch(setStateOperations(multisigOperations))
+    store.dispatch(getMultisigOperationsSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(getMultisigOperationsError(error.message))
+  }),
 })
 
 export const getMultisigOperationsError = (error) => ({
@@ -152,14 +287,118 @@ export const getMultisigOperationsSuccess = (success) => ({
 
 export const getTask = (colonyClient, taskId) => ({
   type: actions.GET_TASK,
-  payload: taskActions.getTaskExtended(colonyClient, taskId)
-    .then(task => {
-      store.dispatch(setStateTask(task))
-      store.dispatch(getTaskSuccess(true))
+  payload: (async () => {
+
+    // Get task
+    const task = await colonyClient.getTask.call({ taskId })
+
+    // Initialize extended protocol
+    await ecp.init()
+
+    // Get specification from specification hash
+    const specification = await ecp.getTaskSpecification(task.specificationHash)
+
+    // Stop extended protocol
+    await ecp.stop()
+
+    // Declare roles
+    const roles = {}
+
+    // Get evaluator
+    roles.evaluator = await colonyClient.getTaskRole.call({
+      taskId,
+      role: 'EVALUATOR',
     })
-    .catch(error => {
-      store.dispatch(getTaskError(error.message))
-    }),
+
+    // Get manager
+    roles.manager = await colonyClient.getTaskRole.call({
+      taskId,
+      role: 'MANAGER',
+    })
+
+    // Get worker
+    roles.worker = await colonyClient.getTaskRole.call({
+      taskId,
+      role: 'WORKER',
+    })
+
+    // Get pot balance
+    const potBalance = await colonyClient.getFundingPotBalance.call({
+      potId: task.potId,
+      token: colonyClient.tokenClient.contract.address,
+    })
+
+    // Get evaluator payout
+    const evaluatorPayout = await colonyClient.getTaskPayout.call({
+      taskId,
+      role: 'EVALUATOR',
+      token: colonyClient.tokenClient.contract.address,
+    })
+
+    // Get manager payout
+    const managerPayout = await colonyClient.getTaskPayout.call({
+      taskId,
+      role: 'MANAGER',
+      token: colonyClient.tokenClient.contract.address,
+    })
+
+    // Get worker payout
+    const workerPayout = await colonyClient.getTaskPayout.call({
+      taskId,
+      role: 'WORKER',
+      token: colonyClient.tokenClient.contract.address,
+    })
+
+    // Declare deliverable
+    let deliverable = {}
+
+    // Check deliverable hash
+    if (task.deliverableHash) {
+
+      // Initialize extended protocol
+      await ecp.init()
+
+      // Get deliverable from deliverable hash
+      deliverable = await ecp.getTaskDeliverable(task.deliverableHash)
+
+      // Stop extended protocol
+      await ecp.stop()
+
+    } else {
+
+      // Set deliverable message
+      deliverable.message = null
+
+    }
+
+    // Get ratings
+    const ratings = await colonyClient.getTaskWorkRatings.call({ taskId })
+
+    // Return task
+    return {
+      ...task,
+      deliverable,
+      payouts: {
+        evaluator: evaluatorPayout.amount.toNumber(),
+        manager: managerPayout.amount.toNumber(),
+        worker: workerPayout.amount.toNumber(),
+      },
+      pot: {
+        balance: potBalance.balance.toNumber(),
+      },
+      ratings,
+      roles: { ...roles },
+      specification: { ...specification },
+    }
+
+  })()
+  .then(task => {
+    store.dispatch(setStateTask(task))
+    store.dispatch(getTaskSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(getTaskError(error.message))
+  }),
 })
 
 export const getTaskError = (error) => ({
@@ -176,14 +415,54 @@ export const getTaskSuccess = (success) => ({
 
 export const getTasks = (colonyClient) => ({
   type: actions.GET_TASKS,
-  payload: taskActions.getTasks(colonyClient)
-    .then(tasks => {
-      store.dispatch(setStateTasks(tasks))
-      store.dispatch(getTasksSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(getTasksError(error.message))
-    }),
+  payload: (async () => {
+
+    // Set tasks
+    let tasks = []
+
+    // Set task id
+    let taskId = 1
+
+    // Get task count
+    const { count: taskCount } = await colonyClient.getTaskCount.call()
+
+    // Loop through tasks
+    while (taskId <= taskCount) {
+
+      // Get task
+      const task = await colonyClient.getTask.call({ taskId })
+
+      // Initialize extended protocol
+      await ecp.init()
+
+      // Get specification from specification hash
+      const specification = await ecp.getTaskSpecification(task.specificationHash)
+
+      // Stop extended protocol
+      await ecp.stop()
+
+      // Add task to tasks
+      tasks.push({
+        ...task,
+        specification: { ...specification },
+      })
+
+      // Increment task id
+      taskId++
+
+    }
+
+    // Return tasks
+    return tasks
+
+  })()
+  .then(tasks => {
+    store.dispatch(setStateTasks(tasks))
+    store.dispatch(getTasksSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(getTasksError(error.message))
+  }),
 })
 
 export const getTasksError = (error) => ({
@@ -200,14 +479,64 @@ export const getTasksSuccess = (success) => ({
 
 export const removeTaskRole = (colonyClient, taskId, role) => ({
   type: actions.REMOVE_TASK_ROLE,
-  payload: taskActions.removeTaskRole(colonyClient, taskId, role)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(removeTaskRoleSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(removeTaskRoleError(error.message))
-    }),
+  payload: (async () => {
+
+    // Check evaluator role
+    if (role === 'EVALUATOR') {
+
+      // Set evaluator role
+      const operation = await colonyClient.removeTaskEvaluatorRole.startOperation({
+        taskId,
+      })
+
+      // Sign multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+      // Return successful
+      return tx.successful
+
+    }
+
+    // Check worker role
+    if (role === 'WORKER') {
+
+      // Set worker role
+      const operation = await colonyClient.removeTaskWorkerRole.startOperation({
+        taskId,
+      })
+
+      // Handle multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+      // Return successful
+      return tx.successful
+
+    }
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(removeTaskRoleSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(removeTaskRoleError(error.message))
+  }),
 })
 
 export const removeTaskRoleError = (error) => ({
@@ -224,14 +553,44 @@ export const removeTaskRoleSuccess = (success) => ({
 
 export const revealRating = (colonyClient, taskId, role, rating) => ({
   type: actions.REVEAL_RATING,
-  payload: taskActions.revealRating(colonyClient, taskId, role, rating)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(revealRatingSuccess(true))
+  payload: (async () => {
+
+    // Set salt
+    const salt = sha3('secret')
+
+    // Generate secret
+    const { secret } = await colonyClient.generateSecret.call({
+      salt: sha3(salt),
+      value: rating,
     })
-    .catch(error => {
-      store.dispatch(revealRatingError(error.message))
-    }),
+
+    // Reveal task work rating
+    const tx = await colonyClient.revealTaskWorkRating.send({
+      taskId,
+      role,
+      rating,
+      salt,
+    })
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(revealRatingSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(revealRatingError(error.message))
+  }),
 })
 
 export const revealRatingError = (error) => ({
@@ -244,11 +603,11 @@ export const revealRatingSuccess = (success) => ({
   payload: success,
 })
 
-// setStateMultisigOperations
+// setStateOperations
 
-export const setStateMultisigOperations = (multisigOperations) => ({
+export const setStateOperations = (operations) => ({
   type: actions.SET_STATE_MULTISIG_OPERATIONS,
-  payload: multisigOperations,
+  payload: operations,
 })
 
 // setStateTask
@@ -256,13 +615,6 @@ export const setStateMultisigOperations = (multisigOperations) => ({
 export const setStateTask = (task) => ({
   type: actions.SET_STATE_TASK,
   payload: task,
-})
-
-// setStateTaskCount
-
-export const setStateTaskCount = (taskCount) => ({
-  type: actions.SET_STATE_TASK_COUNT,
-  payload: taskCount,
 })
 
 // setStateTasks
@@ -276,14 +628,119 @@ export const setStateTasks = (tasks) => ({
 
 export const setTaskDetails = (colonyClient, taskId, details) => ({
   type: actions.SET_TASK_DETAILS,
-  payload: taskActions.setTaskDetails(colonyClient, taskId, details)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(setTaskDetailsSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(setTaskDetailsError(error.message))
-    }),
+  payload: (async () => {
+
+    // Check domain id
+    if (details.domainId) {
+
+      // Format domain id
+      const domainId = details.domainId.toNumber()
+
+      // Set task domain
+      const operation = await colonyClient.setTaskDomain.startOperation({
+        taskId,
+        domainId,
+      })
+
+      // Handle multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Check due date
+    if (details.dueDate) {
+
+      // Format due date
+      const dueDate = new Date(details.dueDate)
+
+      // Set task due date
+      const operation = await colonyClient.setTaskDueDate.startOperation({ taskId, dueDate })
+
+      // Handle multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Check skill id
+    if (details.skillId) {
+
+      // Format skillId
+      const skillId = Number(details.skillId)
+
+      // Set task skill
+      const operation = await colonyClient.setTaskSkill.startOperation({ taskId, skillId })
+
+      // Handle multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Check specification
+    if (details.specification) {
+
+      // Initialize extended protocol
+      await ecp.init()
+
+      // Create specification hash
+      const specificationHash = await ecp.saveTaskSpecification(specification)
+
+      // Stop extended protocol
+      await ecp.stop()
+
+      // start set task brief operation
+      const operation = await colonyClient.setTaskBrief.startOperation({
+        taskId,
+        specificationHash,
+      })
+
+      // Handle multisignature operation
+      const tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // return successful
+    return true
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(setTaskDetailsSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(setTaskDetailsError(error.message))
+  }),
 })
 
 export const setTaskDetailsError = (error) => ({
@@ -300,14 +757,36 @@ export const setTaskDetailsSuccess = (success) => ({
 
 export const setTaskPayouts = (colonyClient, taskId, managerAmount, evaluatorAmount, workerAmount) => ({
   type: actions.SET_TASK_PAYOUTS,
-  payload: taskActions.setTaskPayouts(colonyClient, taskId, managerAmount, evaluatorAmount, workerAmount)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(setTaskPayoutsSuccess(true))
+  payload: (async () => {
+
+    // Set task payouts
+    const tx = await colonyClient.setAllTaskPayouts.send({
+      taskId,
+      token: colonyClient.tokenClient.contract.address,
+      managerAmount: new BN(managerAmount),
+      evaluatorAmount: new BN(evaluatorAmount),
+      workerAmount: new BN(workerAmount),
     })
-    .catch(error => {
-      store.dispatch(setTaskPayoutsError(error.message))
-    }),
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(setTaskPayoutsSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(setTaskPayoutsError(error.message))
+  }),
 })
 
 export const setTaskPayoutsError = (error) => ({
@@ -324,14 +803,88 @@ export const setTaskPayoutsSuccess = (success) => ({
 
 export const setTaskRole = (colonyClient, taskId, role, user) => ({
   type: actions.SET_TASK_ROLE,
-  payload: taskActions.setTaskRole(colonyClient, taskId, role, user)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(setTaskRoleSuccess(true))
-    })
-    .catch(error => {
-      store.dispatch(setTaskRoleError(error.message))
-    }),
+  payload: (async () => {
+
+    // Declare transaction
+    let tx
+
+    // Check manager role
+    if (role === 'MANAGER') {
+
+      // Set task manager role
+      const operation = await colonyClient.setTaskManagerRole.startOperation({
+        taskId,
+        user,
+      })
+
+      // Handle multisignature operation
+      tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Check evaluator role
+    if (role === 'EVALUATOR') {
+
+      // Set task evaluator role
+      const operation = await colonyClient.setTaskEvaluatorRole.startOperation({
+        taskId,
+        user,
+      })
+
+      // Handle multisignature operation
+      tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Check worker role
+    if (role === 'WORKER') {
+
+      // Set worker role
+      const operation = await colonyClient.setTaskWorkerRole.startOperation({
+        taskId,
+        user,
+      })
+
+      // Handle multisignature operation
+      tx = await ms.signOperation(operation);
+
+      // Check unsuccessful
+      if (!tx.successful) {
+
+        // Throw failed transaction error
+        throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+      }
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(setTaskRoleSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(setTaskRoleError(error.message))
+  }),
 })
 
 export const setTaskRoleError = (error) => ({
@@ -348,15 +901,32 @@ export const setTaskRoleSuccess = (success) => ({
 
 export const signTask = (colonyClient, taskId) => ({
   type: actions.SIGN_TASK,
-  payload: taskActions.signTask(colonyClient, taskId)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(getMultisigOperations(colonyClient, taskId))
-      store.dispatch(signTaskSuccess(true))
+  payload: (async () => {
+
+    // Get multisignature operations
+    const operations = await ms.getOperations(colonyClient, taskId)
+
+    // Loop through operations
+    operations.forEach(async operation => {
+
+      // Restore and sign multisignature operation
+      await ms.restoreAndSignOperation(
+        colonyClient,
+        taskId,
+        operations.sender.name,
+      )
+
     })
-    .catch(error => {
-      store.dispatch(signTaskError(error.message))
-    }),
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(getMultisigOperations(colonyClient, taskId))
+    store.dispatch(signTaskSuccess(true))
+  })
+  .catch(error => {
+    store.dispatch(signTaskError(error.message))
+  }),
 })
 
 export const signTaskError = (error) => ({
@@ -373,14 +943,46 @@ export const signTaskSuccess = (success) => ({
 
 export const submitRating = (colonyClient, taskId, role, rating) => ({
   type: actions.SUBMIT_RATING,
-  payload: taskActions.submitRating(colonyClient, taskId, role, rating)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(submitRatingSuccess(true))
+  payload: (async () => {
+
+    // Set salt
+    const salt = sha3('secret')
+
+    // Set value
+    const value = rating
+
+    // Generate secret
+    const { secret } = await colonyClient.generateSecret.call({
+      salt,
+      value,
     })
-    .catch(error => {
-      store.dispatch(submitRatingError(error.message))
-    }),
+
+    // Submit task work rating
+    const tx = await colonyClient.submitTaskWorkRating.send({
+      taskId,
+      role,
+      secret,
+    })
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(submitRatingSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(submitRatingError(error.message))
+  }),
 })
 
 export const submitRatingError = (error) => ({
@@ -397,14 +999,42 @@ export const submitRatingSuccess = (success) => ({
 
 export const submitWork = (colonyClient, taskId, deliverable) => ({
   type: actions.SUBMIT_WORK,
-  payload: taskActions.submitWork(colonyClient, taskId, deliverable)
-    .then(success => {
-      store.dispatch(getTask(colonyClient, taskId))
-      store.dispatch(submitWorkSuccess(true))
+  payload: (async () => {
+
+    // Initialize extended protocol
+    await ecp.init()
+
+    // Create deliverable hash
+    const deliverableHash = await ecp.saveTaskDeliverable(deliverable)
+
+    // Stop extended protocol
+    await ecp.stop()
+
+    // Submit task deliverable
+    const tx = await colonyClient.submitTaskDeliverable.send({
+      taskId,
+      deliverableHash,
     })
-    .catch(error => {
-      store.dispatch(submitWorkError(error.message))
-    }),
+
+    // Check unsuccessful
+    if (!tx.successful) {
+
+      // Throw failed transaction error
+      throw Error ('Transaction Failed: ' + tx.meta.transaction.hash)
+
+    }
+
+    // Return successful
+    return tx.successful
+
+  })()
+  .then(success => {
+    store.dispatch(getTask(colonyClient, taskId))
+    store.dispatch(submitWorkSuccess(success))
+  })
+  .catch(error => {
+    store.dispatch(submitWorkError(error.message))
+  }),
 })
 
 export const submitWorkError = (error) => ({
